Rename metric file names and its references
