+++ src/salicml/metrics/finance/comprovantes_pagamentos.py
