+++ src/salicml/metrics/finance/itens_comuns_e_incomuns_por_segmento.py
