+++ src/salicml/metrics/finance/novos_fornecedores.py
