+++ src/salicml/metrics/finance/preço_itens.py
