+++ src/salicml/metrics/finance/valor_aprovado.py
