+++ src/salicml/metrics/finance/valor_captado.py
